DistUpgradeCache: Ignore BaseMetaPkgs' & obsolete tasks during server upgrade
ubuntu-minimal and ubuntu-base are forced to stay installed via
BaseMetaPkgs / keepBaseMetaPkgs rule during either desktop or server
upgrades. During server upgrades, we try to establish which Tasks are
installed, we do this by iterating all packages and checking if all
packages of a given Task are installed. This is the case for
ubuntu-minimal & ubuntu-standard. But they are not manually-installed
tasks. Instead they are handled with a metapackage only, to allow
alternative implementations of certain components
(i.e. time-daemon). Thus during upgrade we should not force install
those, but instead rely on keeping BaseMetaPkg installed only to
esnure that alternative selections of time-daemons is permitted and is
not changed.

Also, this function might missdetect as obsolete tasks as
installed. For example Task: ubuntu-core should no longer exist, as
the metapackages for it got dropped, but we forgot to remove the
metadata from the archive. If it happens that ubuntu-core appears to
be fully installed, do not try to preserve it on upgrades.

LP: #1872902