Add quirk to check and warn if ROS is installed
Each ROS release is very strict about the versions of Ubuntu it
supports. Operating outside of that guidance can easily lead to upgrade
errors that are hard to debug. Add a quirk to detect if ROS is
installed, and warn the user of this situation before they continue with
the upgrade.

LP: #1611737

Signed-off-by: Kyle Fazzari <kyle@canonical.com>