Add quirk to migrate FKMS overlay to KMS for raspi users
The Raspberry Pi FKMS ("fake" KMS) overlay needs to be replaced with the
KMS overlay ("full" KMS) for correct desktop function under Wayland.
We've been advised there *may* be display incompatibilities with the
camera firmware under this setup, so configuration lines enabling the
camera ("start_x=1") should also be disabled.

Finally, because the camera is disabled, and the GPU is (largely) under
the control of the kernel under this setup, any GPU memory allocation
lines ("gpu_mem=128") should be disabled to revert to the default (64MB)
GPU memory split.

LP: #1923673